requirements.txt
streamlit
pymongo
python-dotenv
requests
boto3
pandas
numpy